+++ ChangeLog.md
@@ -1,3 +1,4 @@
 # Changelog for IdrisTddNotes
 
 ## Unreleased changes
+- 2019-04-20 added notes about eliminating runtime cost of proofs